enemy spawning: - public parameter that controls how often to try and spawn enemies. - uses difficulty to decide how much to speed up spawning over time. - need to uncomment the use of playerpref "difficulty" when merging to   master

+++ Assets/SettingsManager.cs
@@ -36,6 +36,6 @@ public class SettingsManager : MonoBehaviour {
 	}
 
 	public void SaveDifficulty () {
-		PlayerPrefs.SetFloat(difficultyKey, difficultySlider.value);
+		PlayerPrefs.SetInt(difficultyKey, (int)difficultySlider.value);
 	}
 }
